refactor test: rename test module
